requirements.txt
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2023.5.7
charset-normalizer==2.0.4
click==8.0.4
filelock==3.9.0
Flask==2.2.2
fsspec==2023.4.0
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
huggingface-hub==0.15.1
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
llvmlite==0.40.0
MarkupSafe==2.1.1
more-itertools==8.12.0
mpmath==1.2.1
mutagen==1.47.0
networkx==2.8.4
numba==0.57.0
numpy==1.24.3
openai-whisper @ git+https://github.com/openai/whisper.git@5979f03701209bb035a0a466f14131aeb1116cbb
packaging==23.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.21.0
pyparsing==3.0.9
python-dotenv==1.0.1
PyYAML==6.0
regex==2022.7.9
requests==2.32.3
rsa==4.9
safetensors==0.4.5
sympy==1.11.1
tiktoken==0.8.0
tokenizers==0.13.2
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.2.3
websockets==13.1
Werkzeug==2.2.3
yt-dlp==2024.10.22
